v1.0.0?
Main saves and loads password and approved players. Main also stores password, approved players, and the effect used to freeze spectators.

JoinListener intercepts players joining, if there is no password, requests one, if they've supplied a password, ends, and if there is a password but they've not supplied it, demands the password.

Promprs is a list of prompts for use in interacting for the password.